requirements.txt
# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Core dependencies
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Chinese model SDKs
dashscope>=1.10.0

# Vector storage
chromadb>=0.4.0

# Utilities
python-dateutil>=2.8.0
aiohttp>=3.8.0
tiktoken>=0.5.0

# Monitoring
sentry-sdk>=1.30.0

# Dev tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
